Update jupyter examples
